+++ frontend/src/styles/Navbar.css
@@ -1,0 +1,63 @@
+.navbar {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding: 10px 20px;
+    background-color: #59ACE8;
+    box-shadow: 0 2px 4px rgba(0,0,0,0.1);
+  }
+  
+  .logo {
+    width: 300px;
+    height: auto;
+    margin-right: 20px;
+  }
+  
+  .nav-buttons {
+    display: flex;
+    gap: 20px;
+  }
+  
+  .nav-buttons button {
+    background-color: #C5F0FE;
+    border: none;
+    color: black;
+    padding: 10px 20px;
+    border-radius: 40px;
+    cursor: pointer;
+    font-size: 16px;
+    font-weight: bolder;
+    transition: transform 0.2s;
+  }
+  
+  .nav-buttons button:hover {
+    background-color: rgb(40, 144, 255);
+    color: white;
+    transform: scale(1.07);
+    transition: transform 0.2s;
+  }
+  
+  .profile-notifications {
+    display: flex;
+    align-items: center;
+    gap: 10px;
+  }
+  
+  .notifications-button {
+    background-color: transparent;
+    border: none;
+    cursor: pointer;
+    font-size: 16px;
+  }
+  
+  .profile {
+    display: flex;
+    align-items: center;
+  }
+  
+  .profile img {
+    width: 30px;
+    height: 30px;
+    border-radius: 50%;
+  }
+
