+++ frontend/src/styles/Sidebar.css
@@ -1,0 +1,27 @@
+.sidebar {
+    width: 300px;
+    background-color: #D7F0FF;
+    margin-left: 20px;
+    padding: 20px;
+    box-shadow: 0 4px 8px rgba(0,0,0,0.2);
+    border-radius: 40px;
+  }
+  
+  .sidebar a {
+    display: block;
+    padding: 10px;
+    color: black;
+    margin: 5px;
+    text-decoration: none;
+    font-size: 20px;
+    font-weight: bold;
+    border-radius: 40px;
+    transition: all 0.3s ease;
+  }
+  
+  .sidebar a:hover {
+    background-color: #ffffff;
+    border-radius: 40px;
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+  }
+
